fix: define chatFadeIn/chatFadeUp/chatBounce keyframes in chat-history.css
