+++ frontend/src/styles/pages/chat-history.css
@@ -708,3 +708,19 @@
 @media (max-width: 480px) {
   .ch-input-hint { display: none; }
 }
+
+/* ── Animations (self-contained — avoids implicit dependency on chat.css) ──── */
+@keyframes chatFadeIn {
+  from { opacity: 0; }
+  to   { opacity: 1; }
+}
+
+@keyframes chatFadeUp {
+  from { opacity: 0; transform: translateY(8px); }
+  to   { opacity: 1; transform: translateY(0); }
+}
+
+@keyframes chatBounce {
+  0%, 80%, 100% { transform: scale(0.8); opacity: 0.4; }
+  40%            { transform: scale(1.2); opacity: 1;   }
+}
